+++ lightwallet-client-lib/src/main/java/co/electriccoin/lightwallet/client/model/ShieldedProtocolEnum.kt
@@ -4,11 +4,13 @@ import cash.z.wallet.sdk.internal.rpc.Service.ShieldedProtocol
 
 enum class ShieldedProtocolEnum {
     SAPLING,
-    ORCHARD;
+    ORCHARD,
+    IRONWOOD;
 
     fun toProtocol() =
         when (this) {
             SAPLING -> ShieldedProtocol.sapling
             ORCHARD -> ShieldedProtocol.orchard
+            IRONWOOD -> ShieldedProtocol.ironwood
         }
 }