+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/model/TransactionOutput.kt
@@ -7,5 +7,6 @@ data class TransactionOutput(
 enum class TransactionPool {
     TRANSPARENT,
     SAPLING,
-    ORCHARD
+    ORCHARD,
+    IRONWOOD
 }